refactor: extract users_controller concerns for better organization
Split the users_controller into five focused concerns:
- Registration: user sign-up, invitation handling, and exists check
- AiUserManagement: AI user creation, editing, and tool configuration
- ContactManagement: user search and contact preparation logic
- AdminOperations: system admin actions (grant/revoke, lock/unlock, destroy)
- ProfileAndSettings: profile updates, password management, and notifications

This refactoring reduces the controller from ~408 lines to ~9 lines,
improving code organization and maintainability.
